I've added a lot here, I'll try to summarize all of it:
Added a CPU monitor file that grabs the contents of /proc/stat. In addition cpu_monitor.c calculates the
CPU percentage given two sample points.

Added graph.c which renders a sort of line graph that moves from right to left. Wanted to add support for extended
ASCII characters but that got really wonky really fast, so I'm going to punt on that for the forseeable future.

Added threading so that monitoring of CPU/Memory usage can be delegated to different threads.
This essentially establishes UI threads (even though there's some IO taking place on these
threads as well). This enables us to allow keyboard input, which will be necessary when
implementing the process window. We make use of a mutex to break out of the infinite loops
and join our monitoring threads. Before creating the thread the mutex is locked.
On every pass of the infinite loop we check to see if this thread is still locked.
We check this by trying to lock the thread, if the lock succeeds this means the thread
was unlocked, so we can unlock the thread again and return. Currently this mutex is unlocked
on the main thread by any keyboard input, which enables us to leave the child threads and join.

Still need to implement something in the arena library for freeing and allocating regions
manually. But that's a job for another day. It's 1:34AM and I'm dying. I doubt this
explanation even makes any sense.

+++ src/include/window/window.h
@@ -14,12 +14,12 @@ typedef enum _mt_window
 typedef struct _window_data
 {
 	WINDOW *window;
-	unsigned int wHeight, wWidth;
-	unsigned int windowX, windowY;
-	unsigned int paddingTop;
-	unsigned int paddingBottom;
-	unsigned int paddingRight;
-	unsigned int paddingLeft;
+	unsigned short wHeight, wWidth;
+	unsigned short windowX, windowY;
+	unsigned short paddingTop;
+	unsigned short paddingBottom;
+	unsigned short paddingRight;
+	unsigned short paddingLeft;
 	char *windowTitle;
 } WINDOW_DATA;
